requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Mail==0.9.1
PyPDF2==3.0.1
spacy==3.7.2
scikit-learn==1.3.2
pymysql==1.1.0
Werkzeug==3.0.1
SQLAlchemy==2.0.23
mysql-connector-python==8.2.0


